Refactor code into dedicated modules.
